Minor consistency changes
To get full points for passoff we need to be consistent. These are the
styles I'm enforcing:

Use "For each" loops whenever possible...
for(Item i : container) rather than
for(int i = 0; i < container.size(); i++)

Also use braces on everything, even on single if statements, else
statements, etc... (I don't like this style but she complained about
it). I guess as long as we are consistent than we can use either way
but for now it's the way it is.